backend/requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24.0
Pillow>=10.0.0
dashscope>=1.13.3
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
# psycopg2-binary>=2.9.9 # Replaced with MySQL driver
PyMySQL>=1.0.0           # Add PyMySQL driver
python-dotenv>=1.0.0
Flask-SQLAlchemy>=3.0.0  # Add Flask-SQLAlchemy
Flask-Migrate>=4.0.0     # Add Flask-Migrate
alembic>=1.9.0
mysqlclient>=2.1.0
